# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "utopia API",
        "description": "API documentation for utopia ",
        "contact": {
            "email": "info@utopia.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gestion.utopia.madrid/api/",
            "description": "utopia API"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Mostrar categorias",
                "operationId": "31d7eb1fb499fb981765e01717f25c7c",
                "responses": {
                    "200": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "categorias"
                ],
                "summary": "Crear categoria",
                "operationId": "9d672c849a554dcdeb0c9328926d679e",
                "requestBody": {
                    "description": "Post categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "idsector"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    },
                                    "idsector": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "informacion": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
        "/tags": {
            "get": {
                "tags": [
                    "tags"
                ],
                "summary": "Mostrar tags",
                "operationId": "d471af8752732894db4da27409f6e403",
                "responses": {
                    "200": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "tags"
                ],
                "summary": "Crear tag",
                "operationId": "15f49781c05eff0dc6133603d130784b",
                "requestBody": {
                    "description": "Post tag",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}